+++ actstream/managers.py
@@ -8,6 +8,7 @@ from django.contrib.auth import get_user_model
 from actstream.gfk import GFKManager
 from actstream.decorators import stream
 from actstream.registry import check
+from actstream.settings import get_follow_model
 
 
 class ActionManager(GFKManager):
@@ -99,7 +100,7 @@ class ActionManager(GFKManager):
                 actor_object_id=obj.pk
             )
 
-        follows = apps.get_model('actstream', 'follow').objects.filter(user=obj)
+        follows = get_follow_model().objects.filter(user=obj)
         if follow_flag:
             follows = follows.filter(flag=follow_flag)