+++ actstream/settings.py
@@ -1,3 +1,4 @@
+from django.apps import apps
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 
@@ -48,3 +49,27 @@ if USE_DRF:
         DRF_SETTINGS[item] = {
             label.lower(): obj for label, obj in DRF_SETTINGS[item].items()
         }
+
+
+def get_swappable_model(model):
+    model_lookup = getattr(settings, 'ACTSTREAM_%s_MODEL' % model.upper(), 'actstream.%s' % model)
+    try:
+        return apps.get_model(model_lookup, require_ready=False)
+    except ValueError:
+        raise ImproperlyConfigured(
+            "%s must be of the form 'app_label.model_name'" % model_lookup
+        )
+    except LookupError:
+        raise ImproperlyConfigured(
+            "Model '%s' has not been installed" % model_lookup
+        )
+
+
+def get_follow_model():
+    """Return the Follow model that is active."""
+    return get_swappable_model('Follow')
+
+
+def get_action_model():
+    """Return the Action model that is active."""
+    return get_swappable_model('Action')