+++ actstream/tests/test_apps.py
@@ -2,6 +2,8 @@ from unittest import TestCase
 
 from django.apps.registry import apps
 
+from actstream.settings import get_action_model
+
 
 class ActstreamConfigTestCase(TestCase):
 
@@ -10,8 +12,7 @@ class ActstreamConfigTestCase(TestCase):
         actstream_config.ready()
         actstream_config.ready()
 
-        from actstream.models import Action
-        data_fields = [field for field in Action._meta.fields if field.name == 'data']
+        data_fields = [field for field in get_action_model()._meta.fields if field.name == 'data']
         self.assertEqual(
             len(data_fields),
             1