+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 2.0.1
 -----
 
+  - Added support for custom Action and Follow models
   - Added support for Django 5.0 and 5.1
   - Added support for Python 3.10, 3.11 and 3.12
   - Updated GitHub actions